+++ tests/test_manage_settings_token_budget.py
@@ -3,7 +3,7 @@ import asyncio
 import json
 
 import src.settings as settings_mod
-from src.tool_implementations import do_manage_settings
+from src.agent_tools.admin_tools import do_manage_settings
 
 
 def test_set_token_budget_is_not_refused_as_secret(monkeypatch):
